[#105] Refactor: add SelectableWorkoutDashboard and apply to SelectWorkout, EditWorkout
