+++ src/pages/EditWorkout/index.tsx
@@ -2,9 +2,9 @@ import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import { useNavigate } from 'react-router-dom';
 
 import { SaveButton } from '@/components/buttons/SaveButton';
-import { SelectableWorkoutCard } from '@/components/cards/SelectableWorkoutCard';
 import { Headline } from '@/components/Headline';
-import { ROADMAP_DATA, ROUTE_PATH } from '@/constants';
+import { SelectableWorkoutDashboard } from '@/components/SelectableWorkoutDashboard';
+import { ROUTE_PATH } from '@/constants';
 import { usePostWorkoutData } from '@/lib/react-query/useWorkoutData';
 import { accessTokenAtom } from '@/stores/atoms/accessTokenAtom';
 import { modalTypeAtom } from '@/stores/atoms/modalTypeAtom';
@@ -36,29 +36,7 @@ export const EditWorkout = () => {
         <span className="font-extrabold">{nickname}</span>님,
       </Headline>
       <p className="pb-5 text-xs text-[#D9D9D9]">{TEXT_CONTENTS.description}</p>
-      <section className="bg-[#1E1E1E] py-4">
-        <div className="flex w-full justify-center">
-          <p className="pb-1 pt-2 font-bold text-primary">{`현재 가능한 풀업 동작 : ${workoutData.size}`}</p>
-        </div>
-        <div className="grid grid-cols-4 gap-x-3.5 gap-y-6 p-5">
-          {ROADMAP_DATA.map(({ id, name, imageSrc, color }) => {
-            const isCardSelected = workoutData.has(id);
-
-            return (
-              <SelectableWorkoutCard
-                key={id}
-                id={id}
-                name={name}
-                imageSrc={imageSrc}
-                color={color}
-                width="4.75rem"
-                height="6.375rem"
-                isActive={isCardSelected}
-              />
-            );
-          })}
-        </div>
-      </section>
+      <SelectableWorkoutDashboard />
       <div className="pt-5">
         <SaveButton
           isActive
